chore: move palette overrides to mocks dir
