+++ dotcom-rendering/.storybook/mocks/paletteDeclarations.ts
